feat(reactor): renamed LLM2 to LLM and existing LLM to LLM2 (marked as deprecated)
